ecommAPI/requirements.txt
astroid
autopep8
backports.functools-lru-cache
certifi
chardet
configparser
Django
django-cors-headers
django-markdown-deux
djangorestframework
djangorestframework-jwt
django-extra-fields
enum34
isort
lazy-object-proxy
markdown2
mccabe
Pillow
psycopg2
pycodestyle
PyJWT
pylint
pytz
requests
singledispatch
six
urllib3
wrapt
